NDRS-482: Move `block_executor::Event` to separate file.
